Revert "Extract TypeformWidget as a separate component"
This reverts commit c5313d5f6465011e1e2e4f90a5b700c5de46753f.
