+++ features/landing/LandingView.tsx
@@ -15,11 +15,11 @@ import { staticFilesRuntimeUrl } from 'helpers/staticPaths'
 import { Trans, useTranslation } from 'next-i18next'
 import getConfig from 'next/config'
 import React, { ComponentProps, useCallback } from 'react'
+import { useEffect } from 'react'
 import { Box, Button, Card, Flex, Grid, Heading, Image, SxStyleProp, Text } from 'theme-ui'
 import { fadeInAnimation, slideInAnimation } from 'theme/animations'
 
 import { FeaturedIlks, FeaturedIlksPlaceholder } from './FeaturedIlks'
-import { TypeformWidget } from './TypeformWidget'
 
 const {
   publicRuntimeConfig: { apiHost },
@@ -212,6 +212,29 @@ function LandingCards() {
       <LandingCard href={`${apiHost}/daiwallet`} cardKey="dai" />
       <LandingCard href="/support" cardKey="faq" />
     </Grid>
+  )
+}
+
+function TypeformWidget() {
+  const scriptId = 'typeform-widget'
+
+  useEffect(() => {
+    if (document.querySelector(`script#${scriptId}`) !== null) {
+      return
+    }
+    const script = document.createElement('script')
+    script.id = scriptId
+    script.src = '//embed.typeform.com/next/embed.js'
+    document.body.append(script)
+  })
+
+  return (
+    <div
+      dangerouslySetInnerHTML={{
+        __html:
+          '<div data-tf-sidetab="y51linBb" data-tf-width="320" data-tf-height="400" data-tf-custom-icon="https://images.typeform.com/images/VAuKfRAjPyBb" data-tf-button-color="#575CFE" data-tf-button-text="Feedback" style="all:unset;"></div>',
+      }}
+    />
   )
 }
